requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pytest>=8.0.0
pytest-cov>=4.0.0
requests>=2.31.0
playwright>=1.40.0
psycopg2-binary>=2.9.0
openai-whisper>=20231117
pyannote.audio>=3.1.0
